refactored layout, deleted redundant wrapper components, made sticky footer
