+++ src/components/Footer/index.js
@@ -2,11 +2,11 @@ import React from 'react';
 import './styles.scss';
 
 const Footer = (props) => {
-	return (
-		<footer className="footer">
-			<div className="wrap">@ Techintos 2021</div>
-		</footer>
-	);
+  return (
+    <footer className="footer">
+      <div className="wrap">@ Techintos 2021</div>
+    </footer>
+  );
 };
 
 export default Footer;
