+++ src/components/Header/index.js
@@ -5,43 +5,43 @@ import { auth } from './../../firebase/utils';
 import Logo from './../../assets/logo.jpg';
 
 const Header = (props) => {
-	const { currentUser } = props;
-	return (
-		<header className="header">
-			<div className="wrap">
-				<div className="logo">
-					<Link to="/">
-						<img src={Logo} alt="logo" />
-					</Link>
-				</div>
+  const { currentUser } = props;
+  return (
+    <header className="header">
+      <div className="wrap">
+        <div className="logo">
+          <Link to="/">
+            <img src={Logo} alt="logo" />
+          </Link>
+        </div>
 
-				<div className="callToActions">
-					{currentUser && (
-						<ul>
-							<li>
-								<span onClick={() => auth.signOut()}>LogOut</span>
-							</li>
-						</ul>
-					)}
+        <div className="callToActions">
+          {currentUser && (
+            <ul>
+              <li>
+                <span onClick={() => auth.signOut()}>LogOut</span>
+              </li>
+            </ul>
+          )}
 
-					{!currentUser && (
-						<ul>
-							<li>
-								<Link to="/registration">Register</Link>
-							</li>
-							<li>
-								<Link to="/login">Login</Link>
-							</li>
-						</ul>
-					)}
-				</div>
-			</div>
-		</header>
-	);
+          {!currentUser && (
+            <ul>
+              <li>
+                <Link to="/signup">Register</Link>
+              </li>
+              <li>
+                <Link to="/login">Login</Link>
+              </li>
+            </ul>
+          )}
+        </div>
+      </div>
+    </header>
+  );
 };
 
 Header.defaultProps = {
-	currentUser: null
+  currentUser: null,
 };
 
 export default Header;
