+++ src/default.scss
@@ -1,45 +1,45 @@
 html {
-	font-size: 10px;
+  font-size: 10px;
 }
 html,
 body {
-	padding: 0;
-	margin: 0;
+  padding: 0;
+  margin: 0;
 }
 html,
 body,
 #root,
 .app,
 .fullHeight {
-	height: 100%;
+  height: 100%;
 }
 *,
 ::before,
 ::after {
-	-webkit-box-sizing: border-box;
-	-moz-box-sizing: border-box;
-	box-sizing: border-box;
+  -webkit-box-sizing: border-box;
+  -moz-box-sizing: border-box;
+  box-sizing: border-box;
 }
 body {
-	font-family: 'Open Sans', sans-serif;
-	font-size: 1.8rem;
-	font-weight: 400;
-	line-height: 1;
+  font-family: 'Open Sans', sans-serif;
+  font-size: 1.8rem;
+  font-weight: 400;
+  line-height: 1;
 }
 a {
-	text-decoration: none;
-	cursor: pointer;
+  text-decoration: none;
+  cursor: pointer;
 }
 a,
 a:hover {
-	transition: all 4s ease-in-out;
-	-webkit-transition: all 4s ease-in-out;
-	-moz-transition: all 4s ease-in-out;
+  transition: all 4s ease-in-out;
+  -webkit-transition: all 4s ease-in-out;
+  -moz-transition: all 4s ease-in-out;
 }
 .main {
-	width: 100%;
-	max-width: 1449px;
-	height: calc(100% - 6.5rem);
-	padding: 0 10px;
-	margin: 0 auto;
+  width: 100%;
+  max-width: 1449px;
+  height: calc(100% - 6.5rem);
+  padding: 0 10px;
+  margin: 0 auto;
 }
