+++ src/firebase/config.js
@@ -1,8 +1,8 @@
 export const firebaseConfig = {
-	apiKey: 'AIzaSyBaF8y7aD03lQ1_HNUXvx4cfh24276BvnE',
-	authDomain: 'efashion-1b26a.firebaseapp.com',
-	projectId: 'efashion-1b26a',
-	storageBucket: 'efashion-1b26a.appspot.com',
-	messagingSenderId: '208258328542',
-	appId: '1:208258328542:web:664594474b24694e189f4f'
+  apiKey: 'AIzaSyDL4Z8mYFc_JRsRv_NvkkVkUb0aaaLnrrM',
+  authDomain: 'greg-commerce.firebaseapp.com',
+  projectId: 'greg-commerce',
+  storageBucket: 'greg-commerce.appspot.com',
+  messagingSenderId: '827348369515',
+  appId: '1:827348369515:web:31dc18b8aacc68d05df12e',
 };
