--- src/pages/Login/index.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import Signin from '../../components/Signin';
-import './styles.scss';
-
-const Login = (props) => {
-	return (
-		<div>
-			<div>
-				<Signin />
-			</div>
-		</div>
-	);
-};
-
-export default Login;
