+++ src/pages/Recovery/index.js
@@ -1,11 +1,11 @@
 import React from 'react';
 
 const Recovery = (props) => {
-	return (
-		<div>
-			<h1>Email paasword</h1>
-		</div>
-	);
+  return (
+    <div>
+      <h1>Email password</h1>
+    </div>
+  );
 };
 
 export default Recovery;
