--- src/pages/Registration/index.js
@@ -1,14 +1,0 @@
-import React, { Component } from 'react';
-import Signup from './../../components/Signup';
-class Registration extends Component {
-	render() {
-		return (
-			<div>
-				<div>
-					<Signup />
-				</div>
-			</div>
-		);
-	}
-}
-export default Registration;
